Förberedande för studs
Lagt till sköld i playerscript, lite ändringar för dens skull i metoden hurt och lite tillhörande metoder gjorda.
Gjort så att vi har tre olika knappar som heter Attack1(x) Attack2(c) och Block(z) i unitys egna system. Lär göra det lätt att kunna binda om och sådant.

+++ Assets/Scripts/GameManager.cs
@@ -12,7 +12,7 @@ public class GameManager : MonoBehaviour
     [SerializeField]
     public GameObject checkpoint;
 
-    float[] playerStuff = new float[3];
+    int[] playerStuff = new int[3];
     // Start is called before the first frame update
     void Start()
     {
